Add free list to AddressSpace

+++ src/kernel/arch/x86_64/memory/address_space.cpp
@@ -40,7 +40,7 @@ namespace Memory {
         remainingPages = size / 0x1000;
     }
 
-    std::optional<uintptr_t> AddressReservation::allocatePages(int count) {
+    std::optional<uintptr_t> AddressReservation::allocatePages(uint64_t count) {
         
         if (remainingPages < count) {
             return {};
